backend/requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0

# Auth (versiones específicas para compatibilidad)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Validación
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# MQTT
paho-mqtt==2.0.0
certifi>=2024.0.0

# WebSocket
websockets==12.0

# Utilities
python-dateutil==2.8.2
python-json-logger==2.0.7